requirements.txt
blinker==1.8.2
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
Flask==3.0.3
Flask-Cors==5.0.0
idna==3.10
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
packaging==24.1
pluggy==1.5.0
pytest==8.3.3
pytest-mock==3.14.0
python-decouple==3.8
python-dotenv==1.0.1
redis==5.2.0
requests==2.32.3
requests-mock==1.12.1
urllib3==2.2.3
Werkzeug==3.1.1
